refactoring to address prop type issue with info addon
