+++ src/components/data-table/ColumnType.js
@@ -1,0 +1,10 @@
+const ColumnType = {
+    TEXT: Symbol('TEXT'),
+    NUMBER: Symbol('NUMBER'),
+    LINK: Symbol('LINK'),
+    IMAGE: Symbol('IMAGE'),
+    ICON_LINK: Symbol('ICON'),
+    INFO_TEXT: Symbol('INFO_TEXT')
+}
+
+export default ColumnType
